Move component stuff to own file
